Functionality to display number of steps
When search by keyword function is used, the steps(tasks) that are not
selected will be displayed as the total number of steps that are hidden.
This will give an indication to the user to see only the desired task as
well as to know the number of steps between other available/hidden
tasks.